moved core into check
