+++ pkg/collector/check/core/loader.go
