+++ pkg/collector/check/core/loader_test.go
