+++ pkg/collector/check/core/system/memory.go
@@ -2,7 +2,7 @@ package system
 
 import (
 	"github.com/DataDog/datadog-agent/pkg/collector/check"
-	"github.com/DataDog/datadog-agent/pkg/core"
+	"github.com/DataDog/datadog-agent/pkg/collector/check/core"
 	"github.com/op/go-logging"
 	"github.com/shirou/gopsutil/mem"
